[ Many Small Dense Solves ] Using RBF Support Parameter ep=4.080000 Calculating weights (N=25600, n=5, ep=4.080000, hv_k=2, hv_gamma=8.000000e+00) Elapsed time is 11.258257 seconds.
[ Single Large Sparse Solve ]
Calculating weights (N=25600, n=5, ep=4.080000, hv_k=2,
hv_gamma=8.000000e+00)
Elapsed time is 10.059846 seconds.

Thats enough justification to continue with the second approach in
matlab. However, it has other implications: if we want to solve for
weights on the GPU, all we have to do is ONE sparse solve per GPU. And
we know how we can break the mat into subblocks for multiple GPUs. This
is great new. Now I need to figure out if we can solve for weights
reliably with an iterative solver (CG should work since the blocks are
symmetric positive definite). Also, there is a small difference in the
weight values from this large solve. I need to test the accuracy.